refactor: introduce RemoteFrontend service
