+++ packages/@amazeelabs/silverback-search/drupal/silverback_search/silverback_search.info.yml
@@ -3,3 +3,6 @@ type: module
 description: 'Search functionality for Silverback.'
 package: Silverback
 core_version_requirement: ^10 || ^11
+
+dependencies:
+  - silverback_external_preview:silverback_external_preview
