+++ packages/@amazeelabs/silverback-search/drupal/silverback_search/silverback_search.services.yml
@@ -1,0 +1,7 @@
+services:
+  silverback_search.remote_frontend:
+    class: Drupal\silverback_search\Service\RemoteFrontend
+    arguments:
+      - '@silverback_external_preview.external_preview_link'
+      - '@logger.factory'
+      - '@database'
